flask-app/requirements.txt
alembic==1.8.1
bcrypt==3.2.2
blinker==1.5
Brotli==1.0.9
cffi==1.15.1
click==8.1.3
commonmark==0.9.1
dash==2.6.1
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Assets==2.0
Flask-Bcrypt==1.0.1
Flask-Compress==1.12
Flask-Limiter==2.4.6
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
limits==2.6.3
Mako==1.2.1
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.0
packaging==21.3
pandas==1.4.3
plotly==5.10.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
rich==12.4.4
scikit-learn==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.40
tenacity==8.0.1
threadpoolctl==3.1.0
typing_extensions==4.2.0
urllib3==1.26.12
webassets==2.0
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
